refactor(components): refactor empty (#3522)
* refactor(components): refactor empty

* fix: tests
